requirements.txt
asgiref==3.7.2
certifi==2020.12.5
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==3.4.8
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-modeltranslation==0.18.11
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.7
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
transliterate==1.10.2
typing_extensions==4.7.1
urllib3==1.26.4
